Try building from master
Up until now, we were building a specific Git commit on the master
branch. This means we need to push a new change every time we want
to update it.

However, the new Flathub infrastructure allows us to rebuild the app
just with a click of a button.

Building from the tip of master makes this more useful, as each
rebuild will automatically update to a new snapshot without needing
to push any change.