opencolorio: Really clean the CMake files
I previously thought that an opencolorio upgrade had moved the /*.cmake
files into a /cmake/ folder. As a result, I made a change to the cleanup
files in 8b3004bd8baf752897606ea0f7a9f7150c0d6873.

As it turns out, the upgrade didn't move the CMake files, it **added**
the /cmake/ directory in addition to the /*.cmake files.

This commit adds the missing pattern from the cleanup list so that both
are really removed.