Update to 2.81
This also updates some dependencies, usually micro versions of them.

IlmBase is merged into OpenEXR, as upstream now provides a single
tarball.

There is a new dependency, OpenImageDenoise, which only builds on
x86_64. Since it is optional, I'm disabling disabling it on other
architectures, until we figure it out.

Unfortunately, this new release doesn't build on ARM, and upstream
announced they will not support 32 bits architectures any more:

https://lists.blender.org/pipermail/bf-committers/2019-August/050124.html

As a result, we now excude ARM from the list of architectures to build
on Flathub.

Fixes #40