Use the platform ffmpeg
I've been bundling ffmpeg because of its reputation for API/ABI
instability, and to make sure the needed codecs were available.

Upstream ffmpeg apparently improved dramatically on this over the past
few years, to the point where we should look into using the platform
version.

The Freedesktop platform now provides a limited ffmpeg which doesn't
include all the codecs Blender can use, but it also provides an
extension with a full ffmpeg.

This allows distributors to preinstall Blender and the platform without
the nonfree/patented codecs, and users can choose to add the extension
later on.

Fixes #39