Introduce the org.blender.Blender.Codecs extension
We used to bundle our own ffmpeg with all the codecs Blender needs.

But because Endless wanted to preinstall Blender on some of its images,
I made it use the runtime ffmpeg instead, or the one from the
`org.freedesktop.Platform.ffmpeg-full` extension if it is installed.

That runtime extension has H264 support enabled… but doesn't actually
contain any H264 encoder/decoder because it requires the
`org.freedesktop.Platform.openh264` extension for that.

Unfortunately, even when you have both runtime extensions installed,
H264 encoding still doesn't work in Blender because the openh264 encoder
is not good enough. :disappointed:

This commit adds an `org.blender.Blender.Codecs` extension which bundles
the ffmpeg we used to bundle directly in Blender.

Installing it gets all encoding/decoding working again in Blender, but
having it as an extension allows Endless (and other distributors) to
preinstall Blender without worrying about codecs.