Fix external-data-checker annotations (#114)
* Fix flatpak-external-data-checker annotations

The "url" property to the rotating-url checker is meant to be a URL that
serves an HTTP redirect to the latest version of the file. But in these
cases it was pointing to a specific version.

In addition, the "url-template" needs to consistently use the $version
wildcard to allow the new version to be interpolated in.

All these modules are on https://release-monitoring.org/ so we
can use the anitya checker instead, providing a project ID and a
URL template, and adjust the template where needed.

* Add four more flatpak-external-data-checker annotations

There are lots more modules but this is a start!

The previous OpenColorIO annotation was in the wrong block: it should
have been in the definition of a particular source.